renamed application git -> git_manager
