+++ cvat/apps/git_manager/README.md
