+++ cvat/apps/git_manager/__init__.py
