+++ cvat/apps/git_manager/management/__init__.py
