+++ cvat/apps/git_manager/management/commands/__init__.py
