+++ cvat/apps/git_manager/models.py
