+++ cvat/apps/git_manager/urls.py
